Tried to train on all classes but full body but it failed
This was just a long shot with almost no chance of success, but it was
easy to try, so I did it.

This training job failed to learn whisker spot class due to it being on
some pictures and not on others, so the model was getting mixed signals
abou what a whisker spot is. On some pictures its background and on some
its whisker spot.